Update device-restrictions-ios.md
Two notes need to be added for the months of April and May, until the 2005 UI release. The 'Spotlight search to return results from internet' setting is mistakenly duplicated in the console for iOS device restrictions > general. The setting currently only works on supervised devices. With the 2005 UI release, it will work correctly, meaning it work work on device enrolled devices along with supervised devices. The setting will be removed from the 'automated device enrollment' header and will only be left under the less restrictive header.